Cleanup size_t -> int32 warning

+++ src/geom/bvh.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 
 BVHNode::BVHNode(HitableList* list, float t0, float t1)
-	: BVHNode(list->list.data(), list->list.size(), t0, t1)
+	: BVHNode(list->list.data(), (int32)list->list.size(), t0, t1)
 {
 }
